requirements.txt
google-generativeai>=0.8.0
chromadb>=0.5.0
Pillow>=10.0.0
python-magic>=0.4.27
